backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pandas
scikit-learn
pydantic
motor
requests